Switch to Rust stable (#1718)
* Switch to Rust stable

Since we have now removed substrate-wasm-builder dependency we are
abble to switch to rust stable channel.

I needed to update some of the integration tests that dependended
on unstable nightly features.

Fixes: https://github.com/filecoin-project/ref-fvm/issues/1717

* Fix clippy errors

+++ fvm/src/syscalls/gas.rs
@@ -23,5 +23,5 @@ pub fn charge_gas(
 }
 
 pub fn available(context: Context<'_, impl Kernel>) -> Result<u64> {
-    Ok(context.kernel.gas_available().round_down() as u64)
+    Ok(context.kernel.gas_available().round_down())
 }
